requirements.txt
cmake
dlib
opencv-python
face_recognition
mysql-connector-python